Demonstrate authenticated Action API request
This uses the recently-introduced crossorigin= parameter; see T322944
for details, or m3api-examples [1] for another example using this.

Note that the access token is only valid for four hours
($wgOAuth2GrantExpirationInterval), and once it’s expired, the request
will fail CORS, so we can’t even read the error message from the API.
Make sure you log out and back in before trying this out.

[1]: https://github.com/lucaswerkmeister/m3api-examples/tree/main/webapp-clientside-vite-guestbook